Create ledmcp3008.py
The pin1 of the MCP3008 starts from below the half circle you can see in the pin out diagram. MCP3008 ADC has a total of 16 pins out of which 8 pins are for taking the analog input. The analog input pins are from CH0-CH7 (Pins 1-8). On the other side, we have different pins which are as follows

DGND is digital ground pin for the chip.

CS is the chip select pin.

DIN is the data input pin from the Raspberry Pi.

DOUT is the data output pin.

CLK is the clock pin.

AGND is the analog ground pin.

VREF is the analog reference voltage. Connect to 3.3V. You can change it if you want to change the scale.

VDD is the power pin for the chip.


After that, connect the center of the 10k Potentiometer to the CH0 of the MCP3008 and connect the other two ends of Potentiometer to the 3.3V and GND. In the end connect the positive lead of LED which is the longer leg of LED to the GPIO 20 of the Raspberry Pi and then connect the 220 ohm resistor from the negative lead of Raspberry Pi to the GND of Raspberry Pi.

MCP3008	Raspberry Pi
VDD	3.3V
VREF	3.3V
AGND	GND
CLK	GPIO11
DOUT	GPIO9
DIN	GPIO10
CS	GPIO8
DGND	GND